split adapter
