+++ projects/netweave-lending/index.js
@@ -1,0 +1,7 @@
+const { compoundExports2 } = require("../helper/compound");
+
+module.exports = {
+  mode: compoundExports2({
+    comptroller: '0x86112d3176c537B953560EA6fE43f79382E7bffE',
+  })
+}
